Move grid insertion function
